corrected shop page routing
