--- client/src/components/farmer/farmer-list/farmer.list.component.jsx
@@ -1,43 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import { Link } from 'react-router-dom';
-import axios from 'axios';
-
-/* Component Imports */
-import Spinner from 'components/spinner/spinner.component';
-
-/* Styles */
-import './farmer.list.styles.scss';
-
-const FarmerList = ({ match }) => {
-  const [farmers, setFarmers] = useState([]);
-  const [isLoading, setStatusLoading] = useState(true);
-
-  useEffect(() => {
-    axios(`/farmers`)
-      .then(({ data }) => {
-        setFarmers(data.data.farmers);
-        setStatusLoading(false);
-      })
-      .catch((err) => console.log(err.message));
-  }, []);
-
-  return (
-    <section className="farmer-list">
-      {/* First we load a grid with 4 farmers, then we have an option to load more */}
-      <h2>FARMER LIST</h2>
-      {isLoading ? (
-        <Spinner />
-      ) : (
-        <div className="farmer-list__list-container">
-          {farmers.map((farmer) => (
-            <li className="farmer-list__list-container--item">
-              <Link to={`${match.url}/${farmer._id}`}>{farmer.name}</Link>
-            </li>
-          ))}
-        </div>
-      )}
-    </section>
-  );
-};
-
-export default FarmerList;
