--- client/src/components/farmer/farmer-list/farmer.list.styles.scss
@@ -1,15 +1,0 @@
-.farmer-list {
-  height: 30vh;
-  margin-bottom: 5vh;
-  background: #888;
-
-  h2 {
-    text-align: center;
-  }
-
-  &__list {
-    height: 100%;
-    display: flex;
-    justify-content: space-around;
-  }
-}
